requirements.txt
altgraph==0.17
amqp==5.1.1
asgiref==3.4.1
async-timeout==4.0.2
backports.zoneinfo==0.2.1
bcrypt==3.2.2
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
colorlog==6.7.0
commonmark==0.9.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
Deprecated==1.2.13
Django==2.2
django-cors-headers==3.10.1
django-filter==21.1
django-simpleui==2022.6.131
django-timezone-field==4.2.3
djangorestframework==3.13.1
dnspython==2.2.1
dwebsocket==0.5.12
eventlet==0.33.1
flower==1.0.0
future==0.18.2
greenlet==1.1.2
httpie==2.6.0
humanize==3.14.0
idna==3.3
importlib-metadata==4.8.3
importlib-resources==5.4.0
itypes==1.2.0
Jinja2==3.0.3
kombu==5.1.0
macholib==1.14
MarkupSafe==2.0.1
multidict==5.2.0
packaging==21.3
paramiko==2.11.0
pefile==2019.4.18
psycopg2==2.9.3
pycparser==2.21
pycryptodome==3.14.1
Pygments==2.12.0
PyJWT==2.4.0
PyNaCl==1.5.0
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
rich==12.4.4
setuptools==58.0.4
six==1.16.0
SQLAlchemy==1.4.37
sqlparse==0.4.2
typing_extensions==4.1.1
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wheel==0.37.1
wincertstore==0.2
wrapt==1.14.1
zipp==3.6.0
django-filter==21.1
